Fix redundant user fetching
