+++ src/App.jsx
@@ -1,5 +1,4 @@
 import {
-  Outlet,
   Route,
   RouterProvider,
   createBrowserRouter,
@@ -9,27 +8,21 @@ import Reservation from "./routes/Reservation";
 import Root from "./routes/Root";
 import "./App.css";
 import Houses from "./components/House";
-import { useEffect } from "react";
-import { useDispatch } from "react-redux";
-import { getHouses } from "./Redux/HouseSlice";
 import HouseDetails from "./HouseDetails";
 import Signup from "./routes/auth/Signup";
 import Signin from "./routes/auth/Signin";
 import AddHouse from "./components/AddHouse";
 import DeleteHouse from "./components/Delete";
+import MainLayout from "./routes/MainLayout";
 
 export default function App() {
-  const dispatch = useDispatch();
 
-  useEffect(() => {
-    dispatch(getHouses());
-  }, [dispatch]);
   const router = createBrowserRouter(
     createRoutesFromElements(
       <Route
         errorElement={<div>This is the error page</div>}
         path="/"
-        element={<Outlet />}
+        element={<MainLayout />}
       >
         <Route path="/" element={<Root />}>
           {/* Add navigation tabs here */}
